Remove unused exception and V1 artifact classes.
