+++ service/src/main/java/bio/terra/tanagra/app/controller/ConceptSetsV2ApiController.java
@@ -12,8 +12,8 @@ import bio.terra.tanagra.generated.model.*;
 import bio.terra.tanagra.service.*;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
 import bio.terra.tanagra.service.accesscontrol.ResourceIdCollection;
-import bio.terra.tanagra.service.model.ConceptSet;
-import bio.terra.tanagra.service.model.Criteria;
+import bio.terra.tanagra.service.artifact.ConceptSet;
+import bio.terra.tanagra.service.artifact.Criteria;
 import bio.terra.tanagra.service.utils.ToApiConversionUtils;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
