+++ service/src/main/java/bio/terra/tanagra/db/AnnotationDao.java
@@ -5,8 +5,8 @@ import bio.terra.common.db.WriteTransaction;
 import bio.terra.common.exception.MissingRequiredFieldException;
 import bio.terra.common.exception.NotFoundException;
 import bio.terra.tanagra.query.Literal;
-import bio.terra.tanagra.service.model.AnnotationKey;
-import bio.terra.tanagra.service.model.AnnotationValue;
+import bio.terra.tanagra.service.artifact.AnnotationKey;
+import bio.terra.tanagra.service.artifact.AnnotationValue;
 import java.sql.Array;
 import java.util.*;
 import java.util.stream.Collectors;
