+++ service/src/main/java/bio/terra/tanagra/db/CohortDao.java
@@ -7,9 +7,9 @@ import bio.terra.common.exception.NotFoundException;
 import bio.terra.tanagra.exception.SystemException;
 import bio.terra.tanagra.query.filtervariable.BinaryFilterVariable;
 import bio.terra.tanagra.query.filtervariable.BooleanAndOrFilterVariable;
-import bio.terra.tanagra.service.model.Cohort;
-import bio.terra.tanagra.service.model.CohortRevision;
-import bio.terra.tanagra.service.model.Criteria;
+import bio.terra.tanagra.service.artifact.Cohort;
+import bio.terra.tanagra.service.artifact.CohortRevision;
+import bio.terra.tanagra.service.artifact.Criteria;
 import java.sql.Timestamp;
 import java.sql.Types;
 import java.time.Instant;
