+++ service/src/main/java/bio/terra/tanagra/db/ConceptSetDao.java
@@ -4,8 +4,8 @@ import bio.terra.common.db.ReadTransaction;
 import bio.terra.common.db.WriteTransaction;
 import bio.terra.common.exception.NotFoundException;
 import bio.terra.tanagra.exception.SystemException;
-import bio.terra.tanagra.service.model.ConceptSet;
-import bio.terra.tanagra.service.model.Criteria;
+import bio.terra.tanagra.service.artifact.ConceptSet;
+import bio.terra.tanagra.service.artifact.Criteria;
 import java.sql.Timestamp;
 import java.sql.Types;
 import java.time.Instant;
