+++ service/src/main/java/bio/terra/tanagra/db/ReviewDao.java
@@ -8,8 +8,8 @@ import bio.terra.tanagra.exception.SystemException;
 import bio.terra.tanagra.query.Literal;
 import bio.terra.tanagra.query.QueryResult;
 import bio.terra.tanagra.query.RowResult;
-import bio.terra.tanagra.service.model.CohortRevision;
-import bio.terra.tanagra.service.model.Review;
+import bio.terra.tanagra.service.artifact.CohortRevision;
+import bio.terra.tanagra.service.artifact.Review;
 import java.sql.Timestamp;
 import java.time.Instant;
 import java.util.*;
