--- service/src/main/java/bio/terra/tanagra/db/exception/DuplicateCohortException.java
@@ -1,14 +1,0 @@
-package bio.terra.tanagra.db.exception;
-
-import bio.terra.common.exception.BadRequestException;
-
-/** A cohort with this id already exists. */
-public class DuplicateCohortException extends BadRequestException {
-  public DuplicateCohortException(String message) {
-    super(message);
-  }
-
-  public DuplicateCohortException(String message, Throwable cause) {
-    super(message, cause);
-  }
-}
