+++ service/src/main/java/bio/terra/tanagra/service/AnnotationService.java
@@ -7,10 +7,10 @@ import bio.terra.tanagra.exception.SystemException;
 import bio.terra.tanagra.query.Literal;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
 import bio.terra.tanagra.service.accesscontrol.ResourceIdCollection;
-import bio.terra.tanagra.service.model.AnnotationKey;
-import bio.terra.tanagra.service.model.AnnotationValue;
-import bio.terra.tanagra.service.model.Cohort;
-import bio.terra.tanagra.service.model.Review;
+import bio.terra.tanagra.service.artifact.AnnotationKey;
+import bio.terra.tanagra.service.artifact.AnnotationValue;
+import bio.terra.tanagra.service.artifact.Cohort;
+import bio.terra.tanagra.service.artifact.Review;
 import bio.terra.tanagra.service.utils.GcsUtils;
 import bio.terra.tanagra.underlay.Underlay;
 import bio.terra.tanagra.underlay.Underlay.MappingType;
