+++ service/src/main/java/bio/terra/tanagra/service/CohortService.java
@@ -4,8 +4,8 @@ import bio.terra.tanagra.app.configuration.FeatureConfiguration;
 import bio.terra.tanagra.db.CohortDao;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
 import bio.terra.tanagra.service.accesscontrol.ResourceIdCollection;
-import bio.terra.tanagra.service.model.Cohort;
-import bio.terra.tanagra.service.model.CohortRevision;
+import bio.terra.tanagra.service.artifact.Cohort;
+import bio.terra.tanagra.service.artifact.CohortRevision;
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
