+++ service/src/main/java/bio/terra/tanagra/service/ConceptSetService.java
@@ -4,8 +4,8 @@ import bio.terra.tanagra.app.configuration.FeatureConfiguration;
 import bio.terra.tanagra.db.ConceptSetDao;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
 import bio.terra.tanagra.service.accesscontrol.ResourceIdCollection;
-import bio.terra.tanagra.service.model.ConceptSet;
-import bio.terra.tanagra.service.model.Criteria;
+import bio.terra.tanagra.service.artifact.ConceptSet;
+import bio.terra.tanagra.service.artifact.Criteria;
 import java.util.List;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
