+++ service/src/main/java/bio/terra/tanagra/service/FromApiConversionService.java
@@ -7,6 +7,8 @@ import bio.terra.tanagra.query.Literal;
 import bio.terra.tanagra.query.filtervariable.BinaryFilterVariable;
 import bio.terra.tanagra.query.filtervariable.BooleanAndOrFilterVariable;
 import bio.terra.tanagra.query.filtervariable.FunctionFilterVariable;
+import bio.terra.tanagra.service.artifact.Cohort;
+import bio.terra.tanagra.service.artifact.Criteria;
 import bio.terra.tanagra.service.instances.filter.AttributeFilter;
 import bio.terra.tanagra.service.instances.filter.BooleanAndOrFilter;
 import bio.terra.tanagra.service.instances.filter.BooleanNotFilter;
@@ -17,8 +19,6 @@ import bio.terra.tanagra.service.instances.filter.HierarchyParentFilter;
 import bio.terra.tanagra.service.instances.filter.HierarchyRootFilter;
 import bio.terra.tanagra.service.instances.filter.RelationshipFilter;
 import bio.terra.tanagra.service.instances.filter.TextFilter;
-import bio.terra.tanagra.service.model.Cohort;
-import bio.terra.tanagra.service.model.Criteria;
 import bio.terra.tanagra.underlay.*;
 import com.google.common.base.Strings;
 import java.util.Collection;
