+++ service/src/main/java/bio/terra/tanagra/service/ReviewService.java
@@ -11,9 +11,9 @@ import bio.terra.tanagra.query.QueryResult;
 import bio.terra.tanagra.query.TableVariable;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
 import bio.terra.tanagra.service.accesscontrol.ResourceIdCollection;
+import bio.terra.tanagra.service.artifact.Cohort;
+import bio.terra.tanagra.service.artifact.Review;
 import bio.terra.tanagra.service.instances.filter.EntityFilter;
-import bio.terra.tanagra.service.model.Cohort;
-import bio.terra.tanagra.service.model.Review;
 import bio.terra.tanagra.underlay.AttributeMapping;
 import bio.terra.tanagra.underlay.DataPointer;
 import bio.terra.tanagra.underlay.Underlay;
