+++ service/src/main/java/bio/terra/tanagra/service/utils/ToApiConversionUtils.java
@@ -3,11 +3,11 @@ package bio.terra.tanagra.service.utils;
 import bio.terra.tanagra.exception.SystemException;
 import bio.terra.tanagra.generated.model.*;
 import bio.terra.tanagra.query.Literal;
+import bio.terra.tanagra.service.artifact.AnnotationValue;
+import bio.terra.tanagra.service.artifact.Cohort;
+import bio.terra.tanagra.service.artifact.CohortRevision;
+import bio.terra.tanagra.service.artifact.Criteria;
 import bio.terra.tanagra.service.instances.EntityInstanceCount;
-import bio.terra.tanagra.service.model.AnnotationValue;
-import bio.terra.tanagra.service.model.Cohort;
-import bio.terra.tanagra.service.model.CohortRevision;
-import bio.terra.tanagra.service.model.Criteria;
 import bio.terra.tanagra.underlay.Attribute;
 import bio.terra.tanagra.underlay.ValueDisplay;
 import java.util.HashMap;
