+++ service/src/test/java/bio/terra/tanagra/service/ConceptSetServiceTest.java
@@ -7,8 +7,8 @@ import bio.terra.common.exception.NotFoundException;
 import bio.terra.tanagra.app.Main;
 import bio.terra.tanagra.service.accesscontrol.ResourceId;
 import bio.terra.tanagra.service.accesscontrol.ResourceIdCollection;
-import bio.terra.tanagra.service.model.ConceptSet;
-import bio.terra.tanagra.service.model.Study;
+import bio.terra.tanagra.service.artifact.ConceptSet;
+import bio.terra.tanagra.service.artifact.Study;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import org.junit.jupiter.api.AfterEach;
