+++ service/src/test/java/bio/terra/tanagra/service/CriteriaValues.java
@@ -1,6 +1,6 @@
 package bio.terra.tanagra.service;
 
-import bio.terra.tanagra.service.model.Criteria;
+import bio.terra.tanagra.service.artifact.Criteria;
 import java.util.Collections;
 import java.util.List;
 import org.apache.commons.lang3.tuple.Pair;
